requirements.txt
streamlit
google-cloud-bigquery
google-genai